requirements.txt
click>=8.0.0
rich>=12.0.0
aiohttp>=3.8.0
aiofiles>=0.8.0
PyYAML>=6.0
python-dotenv>=0.19.0
gitpython>=3.1.0
asyncio-subprocess